% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{buyandholdSignal}
\alias{buyandholdSignal}
\title{A buy and hold decision function}
\usage{
buyandholdSignal(stocks)
}
\arguments{
\item{stocks}{the list of stocks to decide upon.}
}
\value{
a vector of investment decisions
}
\description{
A buy and hold decision function
}
