% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterStocks}
\alias{clusterStocks}
\title{Cluster a list of stocks into groups}
\usage{
clusterStocks(stocks, k = 10, horizon = 90)
}
\arguments{
\item{stocks}{the list of stocks to operate on typically list of xts objects}

\item{k}{the number of groups to use which defaults to 10}

\item{horizon}{the number of trading days to use for the clustering}
}
\value{
a named class vector
}
\description{
The function clusters stocks into k groups based on a horizon. It
uses k-means clustering on the first PCA component of each stock where
Open, High, Low, Close, Adjusted and Volume goes into the PCA. As such
this clustering should take all sorts of variation into account.
}
\examples{
library(datrader)
mypath <- system.file('extdata', package = 'datrader')
mylist <- loadExistingInstruments(mypath)
getAvailableInstruments(mylist, Sys.Date())
clusterStocks(mylist, k=2, horizon=10)
}
