% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{downloadInstruments}
\alias{downloadInstruments}
\title{Download a list of instrumentr from Yahoo}
\usage{
downloadInstruments(instruments, path, startDate = "2000-01-01")
}
\arguments{
\item{instruments}{the character vector of instruments to download}

\item{path}{the path where to store the final CSV files}

\item{startDate}{the starting date to start the query from}
}
\value{
TRUE if we successfully downloaded all instruments. FALSE otherwise
}
\description{
Download a list of instruments from Yahoo into the specified path which is kept as CSV files
}
\details{
Nothing
}
\examples{
omxs30Symbols<-c('ABB.ST', 'ALFA.ST', 'ASSA-B.ST', 'AZN.ST', 'ATCO-A.ST',
'ATCO-B.ST', 'BOL.ST', 'ELUX-B.ST', 'ERIC-B.ST', 'GETI-B.ST', 'HMB.ST',
'INVE-B.ST', 'LUPE.ST', 'MTG-B.ST', 'NDA-SEK.ST', 'SAND.ST',
'SCA-B.ST', 'SCV-B.ST', 'SEB-A.ST', 'SECU-B.ST', 'SKA-B.ST', 'SKF-B.ST',
'SSAB-A.ST', 'SHB-A.ST', 'SWED-A.ST', 'SWMA.ST', 'TEL2-B.ST', 'TLSN.ST', 'VOLV-B.ST')
\dontrun{downloadInstruments(c('ABB.ST'), '/somewhere/trading')}
}

