% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{extractTradableDates}
\alias{extractTradableDates}
\title{Extract a vector of historical tradable dates}
\usage{
extractTradableDates(instruments)
}
\arguments{
\item{instruments}{the list of instruments to operate on}
}
\value{
a vector of dates where trades happen in one or several instruments
}
\description{
A union of all the traded dates within the list of instruments is returned.
The use case is of course to not try to evaluate a date which has never been
traded on e.g. a weekend.
}
\examples{
a<-1
}
