% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{filterUnwantedInstruments}
\alias{filterUnwantedInstruments}
\title{Filter instruments from the list which are unwanted}
\usage{
filterUnwantedInstruments(stocks, cutoff = 0.05, lastDate = Sys.Date()
  - 1)
}
\arguments{
\item{stocks}{the list of instruments to filter}

\item{cutoff}{the cutoff value of how many NA's are acceptable which is by
default set to 5 per cent}

\item{lastDate}{the date where an instrument have to have data until at least}
}
\value{
the filtered list of instruments
}
\description{
Unwanted instruments are judged by the ratio of NA's compared to total amount
of observations and having data later than the given lastDate.
}
\examples{
a<-1
}
