% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{generateHistoricalPositions}
\alias{generateHistoricalPositions}
\title{Generate historical positions based on a trading strategy}
\usage{
generateHistoricalPositions(x, tstrat)
}
\arguments{
\item{x}{the instrument to generate historical positions in}

\item{tstrat}{the trading strategy to utilize which is based on a single
instrument approach}
}
\value{
a position vector consisting of 0's and 1's for indicating in and out
of position
}
\description{
The trading strategy is based on a single instrument and is allowed to take
multiple extra arguments.
}
\examples{
a<-1
}
