% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{getAvailableInstruments}
\alias{getAvailableInstruments}
\title{Find out which instruments are active at date}
\usage{
getAvailableInstruments(instruments, date)
}
\arguments{
\item{instruments}{the named list of instruments to consider}

\item{date}{the date to use for the query}
}
\value{
a vector of instrument names featuring TRUE or FALSE which indicates
whether or not they existed at the time of date.
}
\description{
Looks at all instruments and writes TRUE if date is newer than the earliest
date for each instrument. This allows us to know if a given instrument
existed at the time of "date".
}
