% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{getFees}
\alias{getFees}
\title{Get the fees for each instrument}
\usage{
getFees(instruments)
}
\arguments{
\item{instruments}{a list of instruments to get the transaction fees for}
}
\value{
a named vector of fees for the instruments
}
\description{
Get the fees for each instrument
}
\examples{
a <- 1
}
