% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{getLastKnownQuantity}
\alias{getLastKnownQuantity}
\title{Get the last known quantity from a list of instruments}
\usage{
getLastKnownQuantity(instruments, f = quantmod::Cl)
}
\arguments{
\item{instruments}{the list of instruments to operate on}

\item{f}{the function that extracts the quantity we would like which defaults
to the closing price.}
}
\value{
a named vector of the last observed quantity for each instrument
}
\description{
Takes a list of instruments and applies the function f to the last known
observation for each instrument.
}
\examples{
library(datrader)
library(quantmod)
library(zoo)
mypath <- system.file('extdata', package = 'datrader')
mylist <- loadExistingInstruments(mypath)
getLastKnownQuantity(mylist, quantmod::Cl)
#' getLastKnownQuantity(mylist, zoo::index)
}
