% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{imputeInstruments}
\alias{imputeInstruments}
\title{Impute the list of instruments}
\usage{
imputeInstruments(stocks)
}
\arguments{
\item{stocks}{the list of instruments to impute NA's in}
}
\value{
a list of the given instruments with NA's imputed
}
\description{
This is a convenience function mapping na.approx to each instrument.
}
\examples{
a<-1
}
