% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{instrumentListToDataFrame}
\alias{instrumentListToDataFrame}
\title{Reduce a list of instruments to a single tibble}
\usage{
instrumentListToDataFrame(x)
}
\arguments{
\item{x}{the list of instruments, which currently all
must be of class xts, to reduce}
}
\value{
a long format tibble
}
\description{
Basically this function reduces a list of instruments
to a single long format tibble in order to create a more
easy to work with format and fascilitate saving as csv
etc.
}
\examples{
a <- 1
}
