% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{instrumentListToTidyquant}
\alias{instrumentListToTidyquant}
\title{Reduce a list of instruments to a single tidyquant tibble}
\usage{
instrumentListToTidyquant(x)
}
\arguments{
\item{x}{the list of instruments, which currently all
must be of class xts, to reduce}
}
\value{
a long format tidyquant tibble
}
\description{
Basically this function reduces a list of instruments
to a single long format tibble in order to create a more
easy to work with format and fascilitate saving as csv
etc.
}
\examples{
library(datrader)
library(quantmod)
library(zoo)
library(tidyquant)
mypath <- system.file('extdata', package = 'datrader')
mylist <- loadExistingInstruments(mypath)
tqdf <- instrumentListToTidyquant(mylist)
mutate(tqdf, date=as.Date(date)) \%>\% filter(date >= "2018-01-01") \%>\%
ggplot(aes(x = date, y = close, group = symbol)) +
geom_candlestick(aes(open = open, high = high, low = low, close = close)) +
theme_tq() + facet_grid(symbol~., scale="free_y")
}
