% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{instrumentListTransToDataFrame}
\alias{instrumentListTransToDataFrame}
\title{Reduce a list of instruments to a single tibble of f}
\usage{
instrumentListTransToDataFrame(x, f = quantmod::Cl)
}
\arguments{
\item{x}{the list of instruments, which currently all
must be of class xts, to reduce}

\item{f}{the function to be applied to each instrument which must create one
single value for each observation in the instrument}
}
\value{
a long format tibble
}
\description{
Basically this function reduces a list of instruments
to a single wide format tibble in order to create a more
easy to work with format and fascilitate saving as csv
etc. It applies function f to each instrument and use that as a column
in the final tibble. The default function applied is the close price
extraction.
}
\examples{
a<-1
}
