% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{is.Portfolio}
\alias{is.Portfolio}
\title{Check if x is a Portfolio}
\usage{
is.Portfolio(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
TRUE if x is a Portfolio FALSE otherwise
}
\description{
Check if x is a Portfolio
}
