% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{loadExistingInstruments}
\alias{loadExistingInstruments}
\title{Load existing instruments from the database into the R environment}
\usage{
loadExistingInstruments(path)
}
\arguments{
\item{path}{the path where all the instruments data are stored as individual CSV files}
}
\value{
the list of xts time series for each instrument
}
\description{
Load existing instruments from the database into the R environment
}
\examples{
a<-1
}

