% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{naInstruments}
\alias{naInstruments}
\title{Get a vector of all instruments containing NA's}
\usage{
naInstruments(instruments)
}
\arguments{
\item{instruments}{the list of instruments to detect NA's in}
}
\value{
a named vector of instruments containing NA's somewhere
}
\description{
Utility function to quickly detect instruments in need of imputing. This
covers all columns of each instrument and reacts to an NA anywhere.
}
\examples{
a<-1
}
