% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{numTrades}
\alias{numTrades}
\title{Calculate number of trades in this position vector}
\usage{
numTrades(p)
}
\arguments{
\item{p}{the position vector indicating whether you are in position or out}
}
\value{
the number of trades made
}
\description{
Takes a vector of 0 and 1 and calculates how many trades are made. A
transition from 0 to 1 or 1 to 0 is calculated as a trade.
}
\examples{
library(datrader)
numTrades(c(0,0,0,0,1,1,1,0,0))
}
