% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{plotHistoricalPositions}
\alias{plotHistoricalPositions}
\title{Plot the historical positions}
\usage{
plotHistoricalPositions(x, pos)
}
\arguments{
\item{x}{the instrument given as an xts object}

\item{pos}{the position vector given by a trading strategy}
}
\value{
a ggplot object representing the plot
}
\description{
This function illustrates the positions generated over time based on a
trading strategy along with the price over time.
}
\examples{
library(datrader)
library(readr)
library(quantmod)
f <- system.file("extdata", "FB.csv", package = "datrader", mustWork = TRUE)
mylist <- loadExistingInstruments(gsub("/FB.csv", "", f))
plotHistoricalPositions(tail(mylist$MSFT, 100), rbinom(100, 1, 0.3))
}
