% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Create a portfolio}
\usage{
Portfolio(holdings, buydates, buyprices, fees)
}
\arguments{
\item{holdings}{the holdings as a named vector of positive integers}

\item{buydates}{a named vector of dates for the transactions}

\item{buyprices}{a named vector of prices for the instruments}

\item{fees}{a named vector of transaction fees for the instruments}
}
\value{
the newly created portfolio
}
\description{
Creates a portfolio by adding the investment, the holdings,
the buy prices as well as the buy dates.
}
\examples{
a <- 1
}
