% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{profit}
\alias{profit}
\alias{profit.Portfolio}
\title{Generic to estimate profit of an object x}
\usage{
profit(x, ...)

\method{profit}{Portfolio}(x, currprices, ...)
}
\arguments{
\item{x}{the object to operate on}

\item{...}{additional arguments to pass to specific method}

\item{currprices}{the prices that represents the market}
}
\value{
the estimated profit as a scalar
}
\description{
Gives and indication of the profit of the object given
a set of arguments.
}
\section{Methods (by class)}{
\itemize{
\item \code{Portfolio}: calculates the profit of a Portfolio
}}

\examples{
library(datrader)
mypath <- system.file('extdata', package = 'datrader')
mylist <- loadExistingInstruments(mypath)

p <- Portfolio(
  holdings=structure(rep(1,length(mylist)), .Names=names(mylist)),
  buydates=sapply(mylist, function(x) tail(index(x), 10)[1]),
  buyprices=sapply(mylist, function(x) tail(Cl(x), 10)[1]),
  fees=structure(rep(1,length(mylist)), .Names=names(mylist)) )

profit(p, sapply(mylist, function(x) tail(Cl(x), 1)))
}
