% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{readInstrument}
\alias{readInstrument}
\title{Read an instrument from a CSV formatted file}
\usage{
readInstrument(fname)
}
\arguments{
\item{fname}{the csv file to load}
}
\value{
a tibble representing the data
}
\description{
The format of the file read mus look like in the details section.
This is also how your resulting tibble will be organized. No other
formats are supported.
}
\details{
"Index","open","high","low","close","volume","adj."
"1999-11-18",32.546494,35.765381,28.612303,31.473534,62546300,27.494957
"1999-11-19",30.71352,30.758226,28.478184,28.880543,15234100,25.229753
"1999-11-22",29.551144,31.473534,28.657009,31.473534,6577800,27.494957
"1999-11-23",30.400572,31.205294,28.612303,28.612303,5975600,24.995413
"1999-11-24",28.701717,29.998211,28.612303,29.372318,4843200,25.659359
}
\examples{
\dontrun{readInstrument("/somewhere/trading/AA.csv")}
}
