% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{sharesToHolding}
\alias{sharesToHolding}
\title{Convert share of investment to actual holding}
\usage{
sharesToHolding(shares, prices, net)
}
\arguments{
\item{shares}{the shares given as percentage for each instrument}

\item{prices}{the price given as a real number for each instrument}

\item{net}{the total amount of money that can be invested}
}
\value{
the holdings as a integer quantity for each instrument
}
\description{
Convert share of investment to actual holding
}
\examples{
library(datrader)
library(quantmod)
mypath <- system.file('extdata', package = 'datrader')
mylist <- loadExistingInstruments(mypath)
rankInstrument <- function(x) tail(momentum(Cl(x), n=90), 1)
selectInstrument <- function(x) rankInstrument(x) > 5
shares <- createPortfolio(mylist, selectInstrument, rankInstrument, topN=3)
prices <- getLastKnownQuantity(mylist, quantmod::Cl)[names(shares)]
sharesToHolding(shares, prices, 10000)
}
