% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{Strategy}
\alias{Strategy}
\title{Type for Investment strategies}
\usage{
Strategy(invest)
}
\arguments{
\item{invest}{a function producing a vector of buying decisions.
See details for more.}
}
\value{
a Strategy instantiated with investment stratey invest.
}
\description{
Type for Investment strategies
}
\examples{
a<-1
}
