% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{updateInstruments}
\alias{updateInstruments}
\title{Update existing database of instruments}
\usage{
updateInstruments(path, startDate = Sys.Date() - 60)
}
\arguments{
\item{path}{the path where all the CSV files are kept}

\item{startDate}{the startDate from when to update with new data where the default is given as today's date - 60 days}
}
\value{
Nothing
}
\description{
Update the existing database of instruments in the specified directory which keep all the data as CSV files.
}
\details{
Nothing.
}
\examples{
\dontrun{updateInstruments("/somewhere/trading")}
}
