% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{evaluateStrategy}
\alias{evaluateStrategy}
\title{Evaluate a strategy on a timeframe and market}
\usage{
evaluateStrategy(instruments, dates, selectInstrument, rankInstrument,
  investFrequency = 30)
}
\arguments{
\item{instruments}{the named list of instruments to consider}

\item{dates}{the timeframe to operate on given as a vector of Date}

\item{selectInstrument}{function that selects whether to invest in an
instrument or not}

\item{rankInstrument}{function that returns a scalar rank for an instrument
higher means better}

\item{investFrequency}{how often to reconsider the portfolio in number of days}
}
\value{
the performance
}
\description{
Evaluate a strategy on a timeframe and market
}
\examples{
a<-1
}
